+++ cmp/zoneGrid/Types.ts
@@ -5,8 +5,8 @@
  * Copyright © 2024 Extremely Heavy Industries Inc.
  */
 
-import {Column, ColumnRenderer, ColumnSortSpec} from '@xh/hoist/cmp/grid';
-import {PersistOptions} from '@xh/hoist/core';
+import {Column, ColumnRenderer, ColumnSortSpec, GridSorterLike} from '@xh/hoist/cmp/grid';
+import {PersistOptions, Some} from '@xh/hoist/core';
 
 export type Zone = 'tl' | 'tr' | 'bl' | 'br';
 
@@ -44,4 +44,11 @@ export interface ZoneGridModelPersistOptions extends PersistOptions {
     persistGrouping?: boolean;
     /** True to include sorting information (default true) */
     persistSort?: boolean;
+}
+
+export interface ZoneGridState {
+    sortBy?: GridSorterLike;
+    groupBy?: Some<string>;
+    version?: number;
+    mappings?: Record<Zone, Some<string | ZoneMapping>>;
 }